enums

+++ esphome/components/ens160/ens160.cpp
@@ -323,7 +323,7 @@ uint16_t ENS160Component::read_u16_le_(uint8_t a_register) {
   return (data >> 8) | (data << 8);
 }
 
-uint16_t ENS160Component::read_u8_(uint8_t a_register) {
+uint8_t ENS160Component::read_u8_(uint8_t a_register) {
   uint8_t data = 0;
   this->read_byte(a_register, &data, 1);
   return (data);
